+++ UI/src/app/modules/core/services/validation-run/validationrun.service.ts
@@ -2,7 +2,7 @@ import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {ValidationrunDto} from './validationrun.dto';
 import {Observable} from 'rxjs';
-import {environment} from '../../../../environments/environment';
+import {environment} from '../../../../../environments/environment';
 
 const publishedValidationRunUrl: string = environment.API_URL + 'api/published-results';
 const customValidationRunUrl: string = environment.API_URL + 'api/my-results';
