+++ UI/src/app/pages/published-validations/published-validations.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {Observable} from 'rxjs';
-import {ValidationrunDto} from '../../modules/validation-result/services/validationrun.dto';
-import {ValidationrunService} from '../../modules/validation-result/services/validationrun.service';
+import {ValidationrunDto} from '../../modules/core/services/validation-run/validationrun.dto';
+import {ValidationrunService} from '../../modules/core/services/validation-run/validationrun.service';
 
 @Component({
   selector: 'qa-published-validations',
